Finished header file restructuring
