+++ include/MAGE/Game.h
@@ -11,18 +11,21 @@
  
  Multiplatform Arcade Game Engine
    Property of Chris Sixsmith.
-       Do not distribute.
+        See licence.txt.
 */
 
 #include "platform.h" // OS stuff
 #include "registry.h"
 #include "hook.h"
-#include "basic.h"
+#include "objBasic.h"
+#include "uiBasic.h"
+
+namespace mage {
 
 class jukebox;
 class mixer;
-class mouseMngr;
-class keyboardMngr;
+class inputMouse;
+class inputKeyboard;
 class gameStateMngr;
 class gameState;
 class resourceMngr;
@@ -31,14 +34,13 @@ class prefabMngr;
 class scriptingEngine;
 
 class uiView;
-class worldView;
+class objView;
 
-class tilemap;
-class player;
+class objTilemap;
 
 class loadingScreen;
 
-class sfTextureResource;
+class resourceTexture;
 
 // The game class is the core of our game.
 // It holds all of the managers and the window.
@@ -86,7 +88,7 @@ public:
 	std::vector<sf::FloatRect> getCollidingBoxes(objBasic* obj);
 	std::vector<std::shared_ptr<objBasic>> findObjectsAt(sf::Vector2f pos);
 	std::vector<std::shared_ptr<uiBasic>> findUiAt(sf::Vector2f pos);
-	std::vector<std::shared_ptr<sfTextureResource>> getKeybindTextures(std::string bName);
+	std::vector<std::shared_ptr<resourceTexture>> getKeybindTextures(std::string bName);
 	void physicsImpulse(sf::Vector2f position, float force, float radius, std::vector<objBasic*> exempt);
 	sf::Vector2f findTextSize(std::string text, unsigned int fontScale = 18);
 
@@ -111,8 +113,8 @@ public:
 
 public:
 	// SUBSYSTEMS
-	std::shared_ptr<mouseMngr> mouse;
-	std::shared_ptr<keyboardMngr> keyboard;
+	std::shared_ptr<inputMouse> mouse;
+	std::shared_ptr<inputKeyboard> keyboard;
 
 	std::shared_ptr<jukebox> music;
 	std::shared_ptr<mixer> sound;
@@ -127,7 +129,7 @@ public:
 	std::shared_ptr<gameState> state;
 
 	std::shared_ptr<uiView> uiCamera;
-	std::shared_ptr<worldView> worldCamera;
+	std::shared_ptr<objView> worldCamera;
 
 	hook<> onGameInitDone;
 	hook<> onGamePreUpdate;
@@ -138,7 +140,7 @@ public:
 	hook<> onGameDone;
 
 	unsigned int mouseMode;
-	std::shared_ptr<sfTextureResource> mouseTexture;
+	std::shared_ptr<resourceTexture> mouseTexture;
 
 	bool showFps;
 	bool debug;
@@ -190,6 +192,8 @@ private:
 };
 
 MAGEDLL Game* theGame();
+
+} // namespace mage
 
 #define DeclareGame(typ) int main(int argc, char *argv[]) {\
 sf::RenderWindow wind;\
