+++ include/MAGE/basic.h
@@ -2,7 +2,7 @@
 #include "StdAfx.h"
 #include "SfmlAfx.h"
 
-#include "animationManager.h"
+#include "animator.h"
 #include "hook.h"
 #include "serializable.h"
 #include "scriptingEngine.h"
@@ -10,13 +10,15 @@
 
 #define CLONEABLE(type) virtual type* clone() const { type* a = new type(*this); a->registerProperties(); return a;}
 
+namespace mage {
+
 class groupBase;
 class prefab;
 
-class sfTextureResource;
-class sfShaderResource;
-class sfSoundBufferResource;
-class sfFontResource;
+class resourceTexture;
+class resourceShader;
+class resourceSoundBuffer;
+class resourceFont;
 class view;
 
 // forward declaration of group:
@@ -90,11 +92,11 @@ public:
 	class MAGEDLL textureData {
 	public:
 		textureData();
-		textureData(std::shared_ptr<sfTextureResource> spriteResource);
-		textureData(std::shared_ptr<sfTextureResource> spriteResource, unsigned int frameWidth, unsigned int frameHeight);
+		textureData(std::shared_ptr<resourceTexture> spriteResource);
+		textureData(std::shared_ptr<resourceTexture> spriteResource, unsigned int frameWidth, unsigned int frameHeight);
 
 	public:
-		std::weak_ptr<sfTextureResource> resource;
+		std::weak_ptr<resourceTexture> resource;
 		sf::Vector2u frameSize;
 		bool defaultFrameSize;
 	};
@@ -153,7 +155,7 @@ public:
 	// helpers
 	sf::Vector2f getCenter() const;
 	const sf::Texture* getTexPointer() const { return sfSprite.getTexture(); }
-	std::shared_ptr<sfTextureResource> getTexResource() const { return getCurrentSprite()->texture.resource.lock(); }
+	std::shared_ptr<resourceTexture> getTexResource() const { return getCurrentSprite()->texture.resource.lock(); }
 	sf::Vector2u getFrameSize() const;
 	sf::Vector2f getRealPosition() const;
 	virtual sf::Vector2f getTexSizeF() const;
@@ -218,106 +220,7 @@ private:
 	bool spriteLoadFailed;
 };
 
-// BASIC WORLD OBJECT
-// ----------------------------------------------------------------------------
-// Supports depth rendering and has collision handling built in (implemented in the Game class)
-// Use for objects that exist in the world.
-
-class MAGEDLL objBasic
-	: public basic
-{
-public:
-	struct MAGEDLL collision {
-		bool hitX, hitY;
-
-		objBasic* main;
-		std::vector<std::shared_ptr<objBasic>> involved;
-	};
-
-public:
-	objBasic(float x, float y, textureData sd = textureData(nullptr));
-	CLONEABLE(objBasic);
-
-	virtual void registerProperties();
-
-	virtual void update(sf::Time elapsed);
-
-	virtual void drawShadows(sf::RenderTarget &target) const;
-	virtual void draw(sf::RenderTarget &target, sf::RenderStates states) const;
-	void drawCollisionBoxes(sf::RenderTarget &target, sf::RenderStates states) const;
-
-	void setPositionC(sf::Vector2f newPosition);
-	void setPositionC(float x, float y);
-
-	objBasic::collision resolveMovement(float byX, float byY);
-
-	void generateCollisionBox(float offL = 0.f, float offR = 0.f, float offT = 0.f, float offB = 0.f);
-	void generateCollisionBoxWithPercentages(float offL = 0.f, float offR = 0.f, float offT = 0.f, float offB = 0.f);
-	sf::FloatRect transformedCollisionBox(int id) const;
-
-	virtual float getDrawBottom();
-
-	virtual bool operator<(objBasic& right)
-	{
-		return (getDrawBottom() < right.getDrawBottom());
-	}
-
-	void playAudio(std::shared_ptr<sfSoundBufferResource> sound, bool replaceSame = false, bool generatePitch = true);
-
-	void pullCamera(std::shared_ptr<view> toPull, float multiplier = 0.1f);
-	void stopPullingCamera();
-public:
-	std::string displayText;
-
-	float drawBottomOffset;
-
-	std::vector<sf::FloatRect> collisionBoxes;
-	
-private:
-	std::weak_ptr<view> pulledCamera;
-	float pulledCameraMultiplier;
-};
-
-// BASIC UI (user interface) OBJECT
-// ----------------------------------------------------------------------------
-// Objects that follow the camera and always render at the same scale.
-// These don't depth render.
-// We're kind of just ignoring the property/clone functionality with this since it's not really meant to be a complex gui system.
-
-class MAGEDLL uiBasic
-	: public basic
-{
-public:
-	CLONEABLE(uiBasic);
-
-	uiBasic(float x, float y, textureData sd = textureData(nullptr));
-
-	virtual void worldDraw(sf::RenderTarget &target, sf::RenderStates states) const;
-
-	virtual void registerProperties();
-
-	virtual bool isMouseOver();
-	bool isOnScreen() const;
-
-	virtual bool operator<(uiBasic& right)
-	{
-		return (getPosition().y > right.getPosition().y);
-	}
-
-	void drawBindPrompt(sf::RenderTarget &target, sf::RenderStates states, int yPosition, std::vector<std::string> keybindNames, std::string caption) const;
-	float drawInfoSimple(sf::RenderTarget &target, sf::RenderStates states, sf::Vector2f screenPos, std::string info) const;
-	float drawInfo(sf::RenderTarget &target, sf::RenderStates states, sf::Vector2f screenPos, std::string info, sf::Color col, bool center = false, unsigned int charSize = 16, float rot = 0.f, std::shared_ptr<sfFontResource> customFont = nullptr) const;
-
-	void drawSelf(sf::RenderTarget &target, sf::RenderStates states, sf::Vector2f screenPos, unsigned int sprite, unsigned int frame = 0, sf::Color col = sf::Color::White) const;
-
-	void bringToFront();
-
-public:
-	hook<const basic*> onWorldDraw;
-};
-
-MAGEDLL bool compareObjects(std::shared_ptr<objBasic> i, std::shared_ptr<objBasic> j);
-MAGEDLL bool compareUi(std::shared_ptr<uiBasic> i, std::shared_ptr<uiBasic> j);
+} // namespace mage
 
 #define DeclareScriptingBasic(type)\
 DeclareScriptingType(type); \
