+++ include/MAGE/objZone.h
@@ -1,11 +1,12 @@
 #pragma once
-#include "basic.h"
+#include "objBasic.h"
+
+namespace mage {
 
 // zone object - keeps track of objects inside it, and can do stuff to them.
-
-class MAGEDLL zone : public objBasic {
+class MAGEDLL objZone : public objBasic {
 public:
-	zone(float x, float y, float sizeX, float sizeY);
+	objZone(float x, float y, float sizeX, float sizeY);
 
 	virtual void update(sf::Time elapsed);
 
@@ -15,7 +16,7 @@ public:
 	bool isInZone(objBasic* obj);
 	bool wasInZone(objBasic* obj);
 
-	CLONEABLE(zone);
+	CLONEABLE(objZone);
 
 public:
 	sf::Vector2f zoneSize;
@@ -26,3 +27,5 @@ public:
 private:
 	std::vector<std::shared_ptr<objBasic>> objectsInZone[2];
 };
+
+} // namespace mage
