+++ include/MAGE/schedules.h
@@ -2,20 +2,14 @@
 #include "StdAfx.h"
 #include "SfmlAfx.h"
 
-#include "easing.h"
 #include "hook.h"
+
+namespace mage {
 
 /*
 The SCHEDULER is used for global animation tasks.
 All tasks CAN take a function that has no arguments and returns a boolean, to verify whether they should keep running, but don't have to
-Tasks take the form of different things:
-- Wait Task: call a function (using a hook<>) after a set amount of time on update. Done using wait(), which returns a task.
-			  (The simplest type of task)
-- Tween Task: call a function (using a hook<float>) every update frame with a varying value. Done using tween(), which returns a task.
-			  (These need an easing function, found in easing.h - if none (nullptr) is given, they will just use Linear)
-			  (If cancelled, a tween will not stop in the middle but skip to the end unless skipOnCancel is set to false)
-- Tick Task: call a function (using a hook<unsigned int>) every set amount of time forever. Done using tick(), which returns a task.
-			  (other than the int argument, just a wait task that doesn't cancel after happening once)
+Tasks take the form of different things.
 */
 
 // you can instantiate the default task but it will go on forever and does nothing.
@@ -46,70 +40,6 @@ private:
 	sf::Time startTime;
 };
 
-// wait task.
-class MAGEDLL waitSchedule : public schedule {
-public:
-	waitSchedule(sf::Time toWait, std::function<void()> callee);
-	waitSchedule(sf::Time toWait, std::function<void()> callee, std::function<bool()> stopCond);
-
-	void run(); // if it is time, call the hook and stop the task.
-
-public:
-	hook<> onWaited;
-	sf::Time length;
-};
-
-// tween task.
-class MAGEDLL tweenSchedule : public schedule {
-public:
-	tweenSchedule(sf::Time toTake, unsigned int phases, easingFunction ef, std::function<void(float)> callee);
-	tweenSchedule(sf::Time toTake, unsigned int phases, easingFunction ef, std::function<void(float)> callee, std::function<bool()> stopCond);
-
-	void start();
-	void run(); // if it is time, call the hook and stop the task.
-	void end(); // if ended, make sure the tween FINISHED before calling the original end() function.
-
-	unsigned int getFrame();
-	sf::Time getFrameLength();
-
-private:
-	void notifyOnTween(float val);
-
-public:
-	hook<float> onTween;
-
-	easingFunction easing;
-
-	sf::Time length;
-	unsigned int numFrames;
-	
-	bool skipOnEnd;
-
-	float tLast;
-};
-
-// tick task.
-class MAGEDLL tickSchedule : public schedule {
-public:
-	tickSchedule(sf::Time toWait, std::function<void(int)> callee);
-	tickSchedule(sf::Time toWait, std::function<void(int)> callee, std::function<bool()> stopCond);
-
-	void start();
-	void run(); // if it is time, call the hook and reset the startTime.
-
-	sf::Time getLastTickTime();
-	sf::Time getTimeSinceLastTick();
-
-	unsigned int getLoopCount();
-public:
-	hook<int> onWaited;
-	sf::Time length;
-
-private:
-	sf::Time lastTickTime;
-	unsigned int timesLooped;
-};
-
 // the manager object
 class MAGEDLL scheduleMngr {
 public:
@@ -126,16 +56,8 @@ public:
 
 	void update(sf::Time elapsed);
 
-	// adder functions
-	std::shared_ptr<waitSchedule> wait(sf::Time toWait, std::function<void()> callee);
-	std::shared_ptr<waitSchedule> wait(sf::Time toWait, std::function<void()> callee, std::function<bool()> stopCond);
-
-	std::shared_ptr<tweenSchedule> tween(sf::Time toTake, unsigned int phases, easingFunction ef, std::function<void(float)> callee);
-	std::shared_ptr<tweenSchedule> tween(sf::Time toTake, unsigned int phases, easingFunction ef, std::function<void(float)> callee, std::function<bool()> stopCond);
-
-	std::shared_ptr<tickSchedule> tick(sf::Time toWait, std::function<void(unsigned int)> callee);
-	std::shared_ptr<tickSchedule> tick(sf::Time toWait, std::function<void(unsigned int)> callee, std::function<bool()> stopCond = nullptr);
-
 private:
 	std::vector<std::shared_ptr<schedule>> tasks;
 };
+
+}
