+++ include/MAGE/view.h
@@ -3,8 +3,9 @@
 // A view draws the entire part of the group it is assigned to that will fit onto the screen.
 // it wraps sf::View to do this.
 // It will do it with a transparent background unless you tell it otherwise
-// It can be told to focus on a specific object and will stop trying to focus on this object when it stops existing.
-// It can have a shader applied to everything it draws (not working yet) as an entire-view feature.
+// It can have a shader applied to everything it draws as an entire-view feature.
+
+// TODO: A view can have render verteces queued. As rendering finishes, it can draw them in.
 
 // The game has 2 it renders with by default - one for UI and one for the world.
 // Both of these are autocreated and configured
@@ -15,6 +16,8 @@
 
 #include "group.h"
 #include "shaders.h"
+
+namespace mage {
 
 class MAGEDLL view : public sf::View, public shadable {
 public:
@@ -63,22 +66,4 @@ public:
 	hook<view*> onRender;
 };
 
-// worldView can see the world objects in a group
-class MAGEDLL worldView : public view {
-public:
-	worldView();
-	worldView(sf::Vector2f size, std::shared_ptr<groupBase> gr = nullptr);
-
-	void render(sf::RenderTarget& target, sf::Color bgCol = sf::Color::Transparent);
-public:
-	bool useStateBounds;
-};
-
-// uiView can see the ui objects in a group
-class MAGEDLL uiView : public view {
-public:
-	uiView();
-	uiView(sf::Vector2f size, std::shared_ptr<groupBase> gr = nullptr);
-
-	void render(sf::RenderTarget& target, sf::Color bgCol = sf::Color::Transparent);
-};
+} // namespace mage
